+++ app/src/main/java/ca/dal/csci3130/quickcash/preferencesmanager/Preferences.java
@@ -4,7 +4,7 @@ public class Preferences implements PreferencesInterface{
 
     private String employeeID;
     private String job;
-    private int salary;
+    private double salary;
     private int startingHour;
     private int startingMinute;
     private int maxDistance;
@@ -26,11 +26,11 @@ public class Preferences implements PreferencesInterface{
         return job;
     }
 
-    public void setSalary(int salary){
+    public void setSalary(double salary){
         this.salary = salary;
     }
 
-    public int getSalary(){
+    public double getSalary(){
         return salary;
     }
